Added Bukkit Implementation * added Minecraft API implementation for Bukkit * done some minor API changes

+++ internalpluginmanager-minecraft-api/src/main/java/com/blockhaus2000/ipm/minecraft/block/meta/BrewingStandBlockMeta.java
@@ -27,7 +27,7 @@ public interface BrewingStandBlockMeta extends BlockMeta, Container {
      * @return The remaining breqing time. Or, if nothing is brewing,
      *         <code>-1</code>.
      */
-    long getRemainingBrewingTime();
+    int getRemainingBrewingTime();
 
     /**
      *
